refactor: greatly simplify batched chat construction and author probability generation.
The prior approach was wildly over complicated for what boils down to string formatting.

In a nutshell, all of the carefully defined substitutions have been replaced with simple truncation + concatenation. In light of these improvements, the default behavior of batch_generate is no longer sequential iteration, and instead is actual batched generation.

There are still some concerns around token boundaries that need to be thoroughly vetted, but all in all, this change should make formatting issues extremely simple to debug compared to the previous method.